requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
requests==2.32.3
python-dotenv==1.0.1
jinja2==3.1.6
apscheduler==3.10.4
python-multipart==0.0.20
